# spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Universal AI Discovery Protocol (DUADP)
  version: "0.1.3"
  description: |
    Decentralized, hybrid-federated protocol for the discovery, validation,
    publishing, and exchange of AI Agents, Skills, Tools, and Marketplaces.
    Built on the Open Standard for Agent Systems (OSSA).
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: OSSA Project
    url: https://openstandardagents.org
servers:
  - url: https://{domain}
    description: Any DUADP node
    variables:
      domain:
        default: example.com
paths:
...
  /api/v1/skills:
    get:
      operationId: listSkills
      summary: List Skills
      description: Returns a paginated list of OSSA-compliant skill manifests.
      tags: [Skills]
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/TrustTier"
        - $ref: "#/components/parameters/Tag"
        - $ref: "#/components/parameters/Runtime"
        - $ref: "#/components/parameters/Action"
        - $ref: "#/components/parameters/ProjectPath"
        - $ref: "#/components/parameters/GroupPath"
        - $ref: "#/components/parameters/Frameworks"
        - $ref: "#/components/parameters/SecurityAttributes"
        - $ref: "#/components/parameters/Federated"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated skills list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillsResponse"
    post:
      operationId: publishSkill
      summary: Publish a Skill
      description: Publish a new skill to this node.
      tags: [Skills]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OssaSkill"
      responses:
        "201":
          description: Skill published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
        "400":
          description: Invalid manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/v1/agents:
    get:
      operationId: listAgents
      summary: List Agents
      description: Returns a paginated list of OSSA-compliant agent manifests.
      tags: [Agents]
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/TrustTier"
        - $ref: "#/components/parameters/Tag"
        - $ref: "#/components/parameters/Runtime"
        - $ref: "#/components/parameters/Action"
        - $ref: "#/components/parameters/ProjectPath"
        - $ref: "#/components/parameters/GroupPath"
        - $ref: "#/components/parameters/Frameworks"
        - $ref: "#/components/parameters/SecurityAttributes"
        - $ref: "#/components/parameters/Federated"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Paginated agents list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentsResponse"
    post:
      operationId: publishAgent
      summary: Publish an Agent
      tags: [Agents]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OssaAgent"
      responses:
        "201":
          description: Agent published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
...
  /api/v1/tools:
    get:
      operationId: listTools
      summary: List Tools
      description: Returns a paginated list of OSSA-compliant tool manifests (MCP, A2A, function-calling).
      tags: [Tools]
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Category"
        - $ref: "#/components/parameters/TrustTier"
        - $ref: "#/components/parameters/Tag"
        - $ref: "#/components/parameters/Federated"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: protocol
          in: query
          schema:
            type: string
            enum: [mcp, a2a, function, rest]
          description: Filter by tool protocol type.
      responses:
        "200":
          description: Paginated tools list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToolsResponse"
    post:
      operationId: publishTool
      summary: Publish a Tool
      tags: [Tools]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OssaTool"
      responses:
        "201":
          description: Tool published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResponse"
...
  /api/v1/control-plane/authorize:
    post:
      operationId: authorizeGitLabContext
      summary: Authorize GitLab runtime context
      description: >-
        Resolves GitLab-aware DUADP tools and skills, evaluates Cedar policy
        for the current MR or pipeline context, and returns the approved catalog,
        allowed hosts, and required reporting checks.
      tags: [Control Plane, GitLab]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GitLabRequestContext"
      responses:
        "200":
          description: Authorization decision for the GitLab context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationDecision"
        "400":
          description: Invalid authorization request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for write operations (publishing, updating, deleting).
    oauth2AuthorizationCode:
      type: oauth2
      description: OAuth2 authorization code flow for user-delegated operations.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.example.com/authorize
          tokenUrl: https://auth.example.com/token
          scopes:
            read: Read access to resources
            write: Publish and update resources
            admin: Administrative operations
    oauth2ClientCredentials:
      type: oauth2
      description: OAuth2 client credentials flow for service-to-service (agent-to-agent) operations.
      flows:
        clientCredentials:
          tokenUrl: https://auth.example.com/token
          scopes:
            read: Read access to resources
            write: Publish and update resources
            federation: Federation peer operations
  parameters:
    Search:
      name: search
      in: query
      schema:
        type: string
      description: Full-text search across name and description.
    Category:
      name: category
      in: query
      schema:
        type: string
      description: Filter by category.
    TrustTier:
      name: trust_tier
      in: query
      schema:
        type: string
        enum: [official, verified-signature, signed, community, experimental]
      description: Filter by trust tier.
    Tag:
      name: tag
      in: query
      schema:
        type: string
      description: Filter by tag (comma-separated for multiple).
    Federated:
      name: federated
      in: query
      schema:
        type: boolean
        default: false
      description: Include results from federated peers.
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number (1-indexed).
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page.
    Runtime:
      name: runtime
      in: query
      schema:
        type: string
        enum: [gitlab]
      description: Runtime-specific query mode.
    Action:
      name: action
      in: query
      schema:
        type: string
      description: GitLab action being evaluated, such as pipeline or merge_request.
    ProjectPath:
      name: project_path
      in: query
      schema:
        type: string
      description: GitLab project path used to scope GitLab-aware catalog results.
    GroupPath:
      name: group_path
      in: query
      schema:
        type: string
      description: GitLab namespace or group path used to scope GitLab-aware catalog results.
    Frameworks:
      name: frameworks
      in: query
      schema:
        type: string
      description: Comma-separated compliance frameworks such as hipaa,fedramp_high.
    SecurityAttributes:
      name: security_attributes
      in: query
      schema:
        type: string
      description: JSON-encoded GitLab security attributes object.
  schemas:
...
    OssaMetadata:
      type: object
      required: [name]
      properties:
        name:
          type: string
        version:
          type: string
        description:
          type: string
        uri:
          type: string
          description: "GAID — agent://namespace/type/name"
        category:
          type: string
        trust_tier:
          type: string
          enum: [official, verified-signature, signed, community, experimental]
        tags:
          type: array
          items:
            type: string
        created:
          type: string
          format: date-time
        updated:
          type: string
          format: date-time
      additionalProperties: true
...
    ResourceIdentity:
      type: object
      required: [did, gaid]
      description: |
        Comprehensive agent/resource identity. Covers DID, DNS binding,
        cryptographic keys, service accounts, provenance, lifecycle,
        operational context, relationships, compliance, and reputation.
      properties:
        did:
          type: string
          description: "W3C DID for this resource (e.g., did:web:example.com:agents:my-agent)."
        gaid:
          type: string
          description: "Global Agent Identifier (agent://namespace/type/name)."
        dns:
          type: object
          properties:
            record:
              type: string
              description: "DNS TXT record name."
            verified:
              type: boolean
        keys:
          type: object
          properties:
            signing:
              type: object
              required: [id, type, public_key_multibase]
              properties:
                id:
                  type: string
                type:
                  type: string
                public_key_multibase:
                  type: string
            encryption:
              type: object
              properties:
                id:
                  type: string
                type:
                  type: string
                public_key_multibase:
                  type: string
            rotation:
              type: object
              properties:
                next_key_hash:
                  type: string
                rotation_policy:
                  type: string
        service_account:
          type: object
          properties:
            id:
              type: string
            type:
              type: string
              enum: [bot, service, system]
            provider:
              type: string
            scopes:
              type: array
              items:
                type: string
            token_endpoint:
              type: string
              format: uri
            client_id:
              type: string
        provenance:
          type: object
          properties:
            creator:
              type: string
            publisher:
              type: string
            created:
              type: string
              format: date-time
            published:
              type: string
              format: date-time
            source_repository:
              type: string
              format: uri
            commit_hash:
              type: string
            build_system:
              type: string
            attestations:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                  uri:
                    type: string
                    format: uri
        lifecycle:
          type: object
          properties:
            status:
              type: string
              enum: [draft, active, suspended, deprecated, revoked]
            activated:
              type: string
              format: date-time
            expires:
              type: [string, "null"]
              format: date-time
            suspended:
              type: [string, "null"]
              format: date-time
            revoked:
              type: [string, "null"]
              format: date-time
            deprecation:
              type: [string, "null"]
              format: date-time
            successor:
              type: [string, "null"]
        operational:
          type: object
          properties:
            endpoint:
              type: string
              format: uri
            protocol:
              type: string
              enum: [mcp, a2a, rest, grpc, websocket]
            transport:
              type: string
              enum: [https, sse, stdio, websocket]
            health_check:
              type: string
              format: uri
            rate_limit:
              type: object
              properties:
                requests_per_minute:
                  type: integer
                concurrent_sessions:
                  type: integer
            availability:
              type: object
              properties:
                sla:
                  type: string
                regions:
                  type: array
                  items:
                    type: string
        relationships:
          type: object
          properties:
            parent_agent:
              type: string
            skills:
              type: array
              items:
                type: string
            tools:
              type: array
              items:
                type: string
            depends_on:
              type: array
              items:
                type: string
            delegates_to:
              type: array
              items:
                type: string
            registered_nodes:
              type: array
              items:
                type: string
        compliance:
          type: object
          properties:
            nist_controls:
              type: array
              items:
                type: string
            safety:
              type: object
              properties:
                human_oversight:
                  type: string
                  enum: [none, optional, recommended, required]
                max_autonomy_level:
                  type: string
                  enum: [autonomous, supervised, human-in-loop, view-only]
                restricted_actions:
                  type: array
                  items:
                    type: string
                safety_policy:
                  type: string
            data_handling:
              type: object
              properties:
                pii_access:
                  type: boolean
                data_retention:
                  type: string
                  enum: [none, session, 30d, 365d, permanent]
                data_residency:
                  type: array
                  items:
                    type: string
                encryption_at_rest:
                  type: boolean
                encryption_in_transit:
                  type: boolean
            audit:
              type: object
              properties:
                log_endpoint:
                  type: string
                  format: uri
                log_format:
                  type: string
                  enum: [OTEL, CEF, JSON, syslog]
                retention_days:
                  type: integer
        reputation:
          type: object
          properties:
            trust_tier:
              type: string
              enum: [official, verified-signature, signed, community, experimental]
            verification_date:
              type: string
              format: date-time
            verified_by:
              type: string
            attestations_count:
              type: integer
            usage_count:
              type: integer
            nodes_registered:
              type: integer
            community_rating:
              type: number
            incidents:
              type: integer

    # ─── New: Agent Card Schema ──────────────────────────────────────
...
    OssaResource:
      type: object
      required: [apiVersion, kind, metadata]
      description: Generic OSSA resource — can be Skill, Agent, Tool, or custom kind.
      properties:
        apiVersion:
          type: string
          pattern: "^ossa/v\\d+\\.\\d+$"
        kind:
          type: string
        metadata:
          $ref: "#/components/schemas/OssaMetadata"
        identity:
          $ref: "#/components/schemas/ResourceIdentity"
          description: "Comprehensive resource identity (DID, keys, service account, lifecycle)."
        spec:
          type: object
        signature:
          $ref: "#/components/schemas/ResourceSignature"
        provenance:
          $ref: "#/components/schemas/ResourceProvenance"
          description: "Supply chain provenance (NIST SP 800-218A)."
        risk:
          $ref: "#/components/schemas/ResourceRisk"
          description: "Risk assessment (NIST AI RMF MAP 5.1)."
        content_hash:
          type: string
          description: "SHA-256 of canonical manifest. Content-addressed dedup."
        endpoints:
          $ref: "#/components/schemas/ProtocolEndpoints"
          description: "Multi-protocol endpoint map (A2A, MCP, OSSA, OpenAPI, REST, etc.)."
        pricing:
          $ref: "#/components/schemas/PricingInfo"
          description: "Cost/pricing metadata."
        sla:
          $ref: "#/components/schemas/SLAInfo"
          description: "Service level agreement."
        status:
          type: string
          enum: [active, deprecated, suspended, sunset]
          description: "Operational status."
        sunset_date:
          type: string
          format: date-time
          description: "Planned decommissioning date (NIST MG-4.1)."
    OssaSkill:
      allOf:
        - $ref: "#/components/schemas/OssaResource"
        - type: object
          properties:
            kind:
              type: string
              const: Skill
    OssaAgent:
      allOf:
        - $ref: "#/components/schemas/OssaResource"
        - type: object
          properties:
            kind:
              type: string
              const: Agent
    OssaTool:
      allOf:
        - $ref: "#/components/schemas/OssaResource"
        - type: object
          properties:
            kind:
              type: string
              const: Tool
    PaginationMeta:
      type: object
      required: [total, page, limit, node_name]
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        node_name:
          type: string
        node_id:
          type: string
        federated:
          type: boolean
        sources:
          type: array
          items:
            type: object
            properties:
              node_id:
                type: string
              node_name:
                type: string
              count:
                type: integer
        next_cursor:
          type: string
          description: Cursor for the next page (cursor-based pagination).
        prev_cursor:
          type: string
          description: Cursor for the previous page (cursor-based pagination).
    SkillsResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OssaSkill"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    AgentsResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OssaAgent"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    ToolsResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/OssaTool"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
...
    PublishResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        resource:
          $ref: "#/components/schemas/OssaResource"
    GitLabSecurityAttributes:
      type: object
      additionalProperties:
        oneOf:
          - type: string
          - type: array
            items:
              type: string
    GitLabToolMetadata:
      type: object
      properties:
        component:
          type: string
        stage:
          type: string
        required_variables:
          type: array
          items:
            type: string
        external_status_check:
          type: string
        external_control:
          type: string
        allowed_hosts:
          type: array
          items:
            type: string
        frameworks:
          type: array
          items:
            type: string
        actions:
          type: array
          items:
            type: string
    GitLabRequestContext:
      type: object
      required: [runtime, action, project_path]
      properties:
        runtime:
          type: string
          enum: [gitlab]
        action:
          type: string
        project_path:
          type: string
        group_path:
          type: string
        project_id:
          oneOf:
            - type: string
            - type: integer
        merge_request_iid:
          type: integer
          nullable: true
        pipeline_id:
          oneOf:
            - type: string
            - type: integer
        sha:
          type: string
        source_branch:
          type: string
        target_branch:
          type: string
        pipeline_source:
          type: string
        labels:
          type: array
          items:
            type: string
        frameworks:
          type: array
          items:
            type: string
        security_attributes:
          $ref: "#/components/schemas/GitLabSecurityAttributes"
        actor:
          type: object
          properties:
            id:
              type: string
            type:
              type: string
              enum: [token, gitlab-job, gitlab-webhook, system]
    AuthorizationDecision:
      type: object
      required: [decision, decision_id, actor, frameworks, approved_tools, approved_skills, allowed_hosts, required_checks, evidence_obligations, cedar]
      properties:
        decision:
          type: string
          enum: [Allow, Deny]
        decision_id:
          type: string
        actor:
          type: string
        frameworks:
          type: array
          items:
            type: string
        approved_tools:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              kind:
                type: string
                enum: [Tool, Skill]
              description:
                type: string
              gitlab:
                $ref: "#/components/schemas/GitLabToolMetadata"
        approved_skills:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              kind:
                type: string
                enum: [Tool, Skill]
              description:
                type: string
              gitlab:
                $ref: "#/components/schemas/GitLabToolMetadata"
        allowed_hosts:
          type: array
          items:
            type: string
        required_checks:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              mode:
                type: string
                enum: [advisory, blocking]
              description:
                type: string
        evidence_obligations:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              framework:
                type: string
              required:
                type: boolean
              description:
                type: string
        cedar:
          type: object
          properties:
            decision:
              type: string
              enum: [Allow, Deny]
            reasons:
              type: array
              items:
                type: string
            errors:
              type: array
              items:
                type: string
            evaluation_ms:
              type: number
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string

    # ─── NIST AI RMF Aligned Types (v0.2.1) ──────────────────
...